+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added UMM option to change the default clothes on the Isekai Protagonist Class.
 ### Changed
+- Changed `Reflect` Special Power to be a toggle ability.
 - Changed `Isekai Protagonist`'s `Friendly Aura` obtained from level 9 -> 10.
 - Changed `Isekai Protagonist` and `Villain`'s '`Quicked-Footed` obtained from level 15 -> 16.
 - Changed `Edge Lord`'s `Fast Movement` obtained from level 7 -> 8.